Fix #1284: rename v1 endpoint classes, Server->RestApiServer
