+++ restapi/src/test/java/io/stargate/web/resources/ColumnResourceTest.java
@@ -12,6 +12,7 @@ import io.stargate.db.schema.Column;
 import io.stargate.db.schema.ImmutableColumn;
 import io.stargate.db.schema.Table;
 import io.stargate.web.models.ColumnDefinition;
+import io.stargate.web.resources.v1.ColumnResource;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
